+++ src/dashboardWebView/components/Media/Media.tsx
@@ -250,8 +250,8 @@ export const Media: React.FunctionComponent<IMediaProps> = (
         )}
 
         <List>
-          {allMedia.map((file) => (
-            <Item key={file.fsPath} media={file} />
+          {allMedia.map((file, idx) => (
+            <Item key={file.fsPath} media={file} index={idx} />
           ))}
         </List>
       </div>